ia-mcps/requirements.txt
# Core
aiohttp>=3.9.0
python-dotenv>=1.0.0

# MCP Server
mcp

# RAG - Vector DB
chromadb>=0.4.0
sentence-transformers>=2.2.0

# File processing
PyMuPDF>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
transformers>=4.35.0
openai-whisper>=20231117

# Dev
ruff>=0.1.0
mypy>=1.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
